requirements.txt
fastapi[standard]
yt-dlp
# openai-whisper
ffmpeg-python
requests
python-dotenv
# torch==2.5.1
